+++ ssh/data/run.sh
@@ -47,13 +47,18 @@ touch /data/.bash_history
 chmod 600 /data/.bash_history
 ln -s -f /data/.bash_history /root/.bash_history
 
-# Build bash_profile from config
-while read -r line ; do
-  echo "$line" >> /root/.bash_profile
-done <<< "$(bashio::config 'bash_profile')"
+# Persist .bash_profile by redirecting .bash_profile to /data
+touch /data/.bash_profile
+chmod 600 /data/.bash_profile
+ln -s -f /data/.bash_profile /root/.bash_profile
 
 # Store token for Hass.io API
-echo "export HASSIO_TOKEN=${HASSIO_TOKEN}" >> /root/.bash_profile
+if grep -q "export HASSIO_TOKEN" /root/.bash_profile;
+  then
+    sed -i "s/^export HASSIO_TOKEN=.*/export HASSIO_TOKEN=${HASSIO_TOKEN}" /root/.bash_profile
+  else
+    echo "export HASSIO_TOKEN=${HASSIO_TOKEN}" >> /root/.bash_profile
+fi
 
 # Start server
 bashio::log.info "Starting SSH daemon..."